backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
transformers==4.36.0
torch==2.1.0
accelerate==0.25.0
sentencepiece==0.1.99
numpy<2.0.0
sacremoses 